sistemato errore JWT e aggiunto caso d'uso abilitaComuneCuratore

+++ digitalterritory/src/main/java/it/unicam/cs/ids/digitalterritory/dto/OsmResponse.java
@@ -13,7 +13,7 @@ public class OsmResponse {
     @JsonProperty("osm_type")
     private String osmType;
     @JsonProperty("osm_id")
-    private int osmId;
+    private long osmId;
     private String lat;
     private String lon;
     private String category;
